+++ runtime-kotlin/runtime-ports/src/main/kotlin/skillbill/ports/concurrency/InterruptSignalPort.kt
@@ -3,9 +3,3 @@ package skillbill.ports.concurrency
 fun interface InterruptSignalPort {
   fun restore()
 }
-
-object JvmInterruptSignalPort : InterruptSignalPort {
-  override fun restore() {
-    Thread.currentThread().interrupt()
-  }
-}